drop the unneeded steamuicoordinator
