+++ client/src/ui/UIManager.ts
@@ -1,55 +1,68 @@
-/**
- * UIManager - Coordinates all UI components and their interactions
- */
-
 import { SteamUIPanel } from './SteamUIPanel'
 import { ProgressDisplay } from './ProgressDisplay'
 import { WebXRUIPanel } from './WebXRUIPanel'
 import { renderTemplate } from '../utils/TemplateEngine'
 import uiErrorTemplate from '../templates/ui/error.html?raw'
+import { EventManager } from '../core/EventManager'
+import { SteamEventTypes } from '../types/InteractionEvents'
+import { steamApi } from '../steam'
 
 export class UIManager {
   private static instance: UIManager | null = null
-  
+
   // Expose UI panels directly instead of using delegation
   public readonly steamUIPanel: SteamUIPanel
   public readonly progressDisplay: ProgressDisplay
   public readonly webxrUIPanel: WebXRUIPanel
-  
+
   constructor() {
     this.steamUIPanel = new SteamUIPanel()
-    
+
     this.progressDisplay = new ProgressDisplay()
-    
+
     this.webxrUIPanel = new WebXRUIPanel()
   }
-  
+
   init(): void {
     this.steamUIPanel.init()
     this.progressDisplay.init()
     this.webxrUIPanel.init()
-    
-    // Show initial UI state
+
     this.showSteamUI()
     this.showControlsHelp()
+
+    const eventManager = EventManager.getInstance()
+    eventManager.registerEventHandler(SteamEventTypes.CacheStats, this.showCacheStats.bind(this))
   }
-  
-  // Overall UI state
+
+  private async showCacheStats(): Promise<void> {
+    try {
+      const cacheManager = steamApi.getCacheManager()
+      const stats = cacheManager.getStats()
+
+      if (stats) {
+        UIManager.getInstance().steamUIPanel.updateCacheStats(stats)
+      }
+    } catch (error) {
+      console.error('Failed to get cache stats:', error)
+    }
+  }
+
   private showSteamUI(): void {
     this.steamUIPanel.show()
   }
-  
+
   private showControlsHelp(): void {
     this.webxrUIPanel.showControlsHelp()
   }
-  
+
   hideLoading(): void {
     const loading = document.getElementById('loading')
     if (loading) {
       loading.classList.add('hidden')
     }
   }
-  
+
   showError(message: string): void {
     const loading = document.getElementById('loading')
     if (loading) {
@@ -57,12 +70,8 @@ export class UIManager {
     }
   }
 
-  /**
-   * Get singleton instance - creates itself when first accessed
-   */
   static getInstance(): UIManager {
     if (!UIManager.instance) {
-      // Self-initialize with no dependencies - truly self-sufficient
       UIManager.instance = new UIManager()
     }
     return UIManager.instance
