+++ client/src/ui/coordinators/index.ts
@@ -2,11 +2,9 @@
  * UI Coordinators - Specialized UI management classes
  * 
  * Exports for the refactored UI coordinator architecture:
- * - SteamUICoordinator: Steam-specific UI workflows  
  * - WebXRUICoordinator: WebXR-specific UI management
  * - SystemUICoordinator: System-level UI (pause menu, performance, settings)
  */
 
-export { SteamUICoordinator } from './SteamUICoordinator'
 export { WebXRUICoordinator } from './WebXRUICoordinator'
 export { SystemUICoordinator } from './SystemUICoordinator'
