--- client/test/unit/ui/coordinators/SteamUICoordinator.test.ts
@@ -1,50 +1,0 @@
-/**
- * Unit tests for SteamUICoordinator
- * Tests constructor and ensures class can be instantiated properly
- */
-
-import { describe, it, expect } from 'vitest'
-import { SteamUICoordinator } from '../../../../src/ui/coordinators/SteamUICoordinator'
-
-describe('SteamUICoordinator Unit Tests', () => {
-  
-  describe('Constructor and Basic Functionality', () => {
-    it('should instantiate without throwing errors', () => {
-      expect(() => {
-        new SteamUICoordinator()
-      }).not.toThrow()
-    })
-
-    it('should have all expected public methods', () => {
-      const coordinator = new SteamUICoordinator()
-      
-      // Verify remaining methods exist
-      expect(typeof coordinator.loadFromCache).toBe('function')
-      expect(typeof coordinator.showCacheStats).toBe('function')
-      expect(typeof coordinator.clearImageCache).toBe('function')
-      expect(typeof coordinator.updateCacheStats).toBe('function')
-    })
-
-    it('should handle method calls without crashing (basic smoke test)', () => {
-      const coordinator = new SteamUICoordinator()
-      
-      // These should not throw even if they don't work properly due to missing DOM/mocks
-      expect(() => {
-        coordinator.loadFromCache('test')
-        coordinator.showCacheStats()
-        coordinator.clearImageCache()
-      }).not.toThrow()
-    })
-  })
-
-  describe('Method Parameter Validation', () => {
-    it('should handle updateCacheStats with valid stats object', () => {
-      const coordinator = new SteamUICoordinator()
-      const mockStats = { totalEntries: 5, cacheHits: 3, cacheMisses: 2 }
-      
-      expect(() => {
-        coordinator.updateCacheStats(mockStats)
-      }).not.toThrow()
-    })
-  })
-})
